% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocYEAR.R
\docType{data}
\name{CoC_2006}
\alias{CoC_2006}
\title{2006 Continuum of Care (CoC) GIS shapefiles and Homeless Counts}
\format{An object of class \code{SpatialPolygonsDataFrame} with 366 rows and 9 columns.}
\source{
Almquist, Helwig and You (2017). Working Paper.

https://www.hudexchange.info/programs/coc/gis-tools/

https://www.hudexchange.info/resource/3031/pit-and-hic-data-since-2007/
}
\usage{
CoC_2006
}
\description{
These GIS shapefiles contain the geographic boundaries and funding status information
for HUD’s Continuum of Care (CoC) areas by Year.
}
\keyword{datasets}

