% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocYEAR.R
\docType{data}
\name{CoC_2011}
\alias{CoC_2011}
\title{2011 Continuum of Care (CoC) GIS shapefiles and Homeless Counts}
\format{An object of class \code{data.frame} with 424 rows and 33 columns.}
\source{
https://www.hudexchange.info/programs/coc/gis-tools/

https://www.hudexchange.info/resource/3031/pit-and-hic-data-since-2007/
}
\usage{
CoC_2011
}
\description{
These GIS shapefiles contain the geographic boundaries and funding status information
for HUD’s Continuum of Care (CoC) areas by Year.
}
\keyword{datasets}

