% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocYEAR.R
\docType{data}
\name{USCountyTotalHomeless}
\alias{USCountyTotalHomeless}
\title{County level homeless counts disaggregated by Density and Shapefiles}
\format{A data frame with 3143 rows:
\describe{
}}
\source{
Almquist, Helwig and You (2017). Working Paper.

http://users.stat.umn.edu/~almquist/software.html

https://www.hudexchange.info/resource/3031/pit-and-hic-data-since-2007/
}
\usage{
USCountyTotalHomeless
}
\description{
Completed county-level homeless counts (with missing values imputed) and selected
demographic variables SF 1 Summary file. HUD provided CoC counts disaggregated
and imputed to align with US County shape files. Includes Homeless counts 200-2017
dissagregated by density.
}
\keyword{datasets}

