% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocYEAR.R
\docType{data}
\name{states}
\alias{states}
\title{US State polygon files}
\format{Spatial Polygon DataFrame:
\describe{
}}
\source{
US Census Bureau
}
\usage{
states
}
\description{
Completed US State Polygon Files
}
\keyword{datasets}

